requirements.txt
Django==2.1.5
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
django-braces==1.13.0
requests-oauthlib==1.1.0
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.0
idna==2.8
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
django-filter